+++ src/scib_metrics/__init__.py
@@ -2,11 +2,11 @@ import logging
 from importlib.metadata import version
 
 from . import utils
-from ._ari_nmi import nmi_ari_cluster_labels_kmeans, nmi_ari_cluster_labels_leiden
 from ._graph_connectivity import graph_connectivity
 from ._isolated_labels import isolated_labels
 from ._kbet import kbet, kbet_per_label
 from ._lisi import clisi_knn, ilisi_knn, lisi_knn
+from ._nmi_ari import nmi_ari_cluster_labels_kmeans, nmi_ari_cluster_labels_leiden
 from ._pcr_comparison import pcr_comparison
 from ._settings import settings
 from ._silhouette import silhouette_batch, silhouette_label
